+++ samples/HelloWorld/Program.cs
@@ -1,9 +1,7 @@
 // Copyright (c) Nate McMaster.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
-using McMaster.Extensions.CommandLineUtils;
 
 public class Program
 {