+++ test/CommandLineUtils.Tests/HelpOptionAttributeTests.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.IO;
-using System.Reflection;
 using System.Text;
 using Xunit;
 using Xunit.Abstractions;